create an analytics.py module for plot and data analytics, move things out from experiment and util
