+++ rl/analytics.py
@@ -1,0 +1,167 @@
+import matplotlib
+import numpy as np
+import pandas as pd
+import platform
+import warnings
+from os import environ
+from rl.util import *
+
+warnings.filterwarnings("ignore", module="matplotlib")
+
+# TODO fix mp breaking on Mac shit,
+# except when running -b with agg backend
+# (no GUI rendered,but saves graphs)
+# set only if it's not MacOS
+if environ.get('CI') or platform.system() == 'Darwin':
+    matplotlib.rcParams['backend'] = 'agg'
+else:
+    matplotlib.rcParams['backend'] = 'TkAgg'
+
+
+class Grapher(object):
+
+    '''
+    Grapher object that belongs to a Session
+    to draw graphs from its data
+    '''
+
+    def __init__(self, session):
+        if not args.plot_graph:
+            return
+        import matplotlib.pyplot as plt
+        plt.rcParams['toolbar'] = 'None'  # mute matplotlib toolbar
+        self.plt = plt
+        self.session = session
+        self.graph_filename = self.session.graph_filename
+        self.subgraphs = {}
+        self.figure = self.plt.figure(facecolor='white', figsize=(8, 9))
+        self.figure.suptitle(wrap_text(self.session.session_id))
+        self.init_figure()
+
+    def init_figure(self):
+        if environ.get('CI'):
+            return
+        # graph 1
+        ax1 = self.figure.add_subplot(
+            311,
+            frame_on=False,
+            title="\n\ntotal rewards per episode",
+            ylabel='total rewards')
+        p1, = ax1.plot([], [])
+        self.subgraphs['total rewards'] = (ax1, p1)
+
+        ax1e = ax1.twinx()
+        ax1e.set_ylabel('exploration rate').set_color('r')
+        ax1e.set_frame_on(False)
+        p1e, = ax1e.plot([], [], 'r')
+        self.subgraphs['e'] = (ax1e, p1e)
+
+        # graph 2
+        ax2 = self.figure.add_subplot(
+            312,
+            frame_on=False,
+            title='mean rewards over last 100 episodes',
+            ylabel='mean rewards')
+        p2, = ax2.plot([], [], 'g')
+        self.subgraphs['mean rewards'] = (ax2, p2)
+
+        # graph 3
+        ax3 = self.figure.add_subplot(
+            313,
+            frame_on=False,
+            title='loss over time, episode',
+            ylabel='loss')
+        p3, = ax3.plot([], [])
+        self.subgraphs['loss'] = (ax3, p3)
+
+        self.plt.tight_layout()  # auto-fix spacing
+        self.plt.ion()  # for live plot
+
+    def plot(self):
+        '''do live plotting'''
+        if not args.plot_graph or environ.get('CI'):
+            return
+        sys_vars = self.session.sys_vars
+        ax1, p1 = self.subgraphs['total rewards']
+        p1.set_ydata(
+            sys_vars['total_rewards_history'])
+        p1.set_xdata(np.arange(len(p1.get_ydata())))
+        ax1.relim()
+        ax1.autoscale_view(tight=True, scalex=True, scaley=True)
+
+        ax1e, p1e = self.subgraphs['e']
+        p1e.set_ydata(
+            sys_vars['explore_history'])
+        p1e.set_xdata(np.arange(len(p1e.get_ydata())))
+        ax1e.relim()
+        ax1e.autoscale_view(tight=True, scalex=True, scaley=True)
+
+        ax2, p2 = self.subgraphs['mean rewards']
+        p2.set_ydata(
+            sys_vars['mean_rewards_history'])
+        p2.set_xdata(np.arange(len(p2.get_ydata())))
+        ax2.relim()
+        ax2.autoscale_view(tight=True, scalex=True, scaley=True)
+
+        ax3, p3 = self.subgraphs['loss']
+        p3.set_ydata(sys_vars['loss'])
+        p3.set_xdata(np.arange(len(p3.get_ydata())))
+        ax3.relim()
+        ax3.autoscale_view(tight=True, scalex=True, scaley=True)
+
+        self.plt.draw()
+        self.plt.pause(0.01)
+        self.save()
+
+    def save(self):
+        '''save graph to filename'''
+        self.figure.savefig(self.graph_filename)
+
+
+def basic_stats(array):
+    '''generate the basic stats for a numerical array'''
+    if not len(array):
+        return None
+    return {
+        'min': np.min(array).astype(float),
+        'max': np.max(array).astype(float),
+        'mean': np.mean(array).astype(float),
+        'std': np.std(array).astype(float),
+    }
+
+
+def analyze_param_space(experiment_data_array_or_prefix_id):
+    '''
+    get all the data from all experiments.run()
+    or read from all data files matching the prefix of experiment_id
+    e.g. usage without running:
+    prefix_id = 'DevCartPole-v0_DQN_LinearMemoryWithForgetting_BoltzmannPolicy_2017-01-15_142810'
+    analyze_param_space(prefix_id)
+    '''
+    if isinstance(experiment_data_array_or_prefix_id, str):
+        experiment_data_array = load_data_array_from_prefix_id(
+            experiment_data_array_or_prefix_id)
+    else:
+        experiment_data_array = experiment_data_array_or_prefix_id
+
+    flat_metrics_array = []
+    for data in experiment_data_array:
+        flat_metrics = flatten_dict(data['summary']['metrics'])
+        flat_metrics.update({'experiment_id': data['experiment_id']})
+        flat_metrics_array.append(flat_metrics)
+
+    metrics_df = pd.DataFrame.from_dict(flat_metrics_array)
+    metrics_df.sort_values(
+        ['mean_rewards_per_epi_stats_mean',
+         'mean_rewards_stats_mean', 'solved_ratio_of_sessions'],
+        ascending=False
+    )
+
+    experiment_id = experiment_data_array[0]['experiment_id']
+    prefix_id = prefix_id_from_experiment_id(experiment_id)
+    param_space_data_filename = './data/{0}/param_space_data_{0}.csv'.format(
+        prefix_id)
+    metrics_df.to_csv(param_space_data_filename, index=False)
+    logger.info(
+        'Param space data saved to {}'.format(param_space_data_filename))
+    return metrics_df
